+++ example/gpio-info.js
@@ -8,7 +8,7 @@ for (gpioNo = Gpio.MIN_GPIO; gpioNo <= Gpio.MAX_GPIO; gpioNo += 1) {
   gpio = new Gpio(gpioNo);
 
   console.log('GPIO ' + gpioNo + ':' +
-    ' mode=' + gpio.getPinMode() +
+    ' mode=' + gpio.getMode() +
     ' level=' + gpio.digitalRead()
   );
 }